move all reducers to reducers folder
1) move all reducers to reducers folder
since all reducers are referenced via rootReducer anyway
