+++ hubmap/frontend/src/middleware/reducers/geneTissueColorsReducer.js
@@ -1,4 +1,4 @@
-import * as Constants from '../commons/constants';
+import * as Constants from '../../commons/constants';
 
 // define DEFAULT state, so it is never "undefined".
 const DEFAULT_STATE = {
@@ -9,11 +9,11 @@ const DEFAULT_STATE = {
 };
 
 /**
- * fill in state object for experiments
+ * fill in the state object for colors
  */
-export default function searchResponseReducer(state = DEFAULT_STATE, action) {
+export default function geneTissueColorResponseReducer(state = DEFAULT_STATE, action) {
   switch (action.type) {
-    case Constants.SEARCH_EXPERIMENTS:
+    case Constants.GET_GENE_TISSUE_COLORS:
       return {
         ...state,
         response: action.response,
@@ -21,6 +21,7 @@ export default function searchResponseReducer(state = DEFAULT_STATE, action) {
         status: action.status,
         error: action.error,
       };
+
     default:
       return state;
   }
