+++ hubmap/frontend/src/middleware/rootReducer.js
@@ -1,8 +1,8 @@
 import { combineReducers } from 'redux';
-import colorsReducer from './colorsReducer';
-import experimentsReducer from './experimentsReducer';
-import geneTissueColorResponseReducer from './geneTissueColorsReducer';
-import searchReducer from './searchReducer';
+import colorsReducer from './reducers/colorsReducer';
+import experimentsReducer from './reducers/experimentsReducer';
+import geneTissueColorResponseReducer from './reducers/geneTissueColorsReducer';
+import searchReducer from './reducers/searchReducer';
 
 export default combineReducers({
   experimentState: experimentsReducer,
